Fix margin issues from 'img-row' element
Apply flex on img-card, changed sizing for img under img-card

The previous commit caused the card to have a definitive value which messed with the overall margin. To fix this, the img-card has specified width and given flex property. The sizing and cropping of the image is handled by width and height inside the img tag.